modules/location_module/requirements.txt
annotated-types==0.6.0
attrs==23.1.0
Authlib==1.3.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
flasgger==0.9.7.1
Flask==3.0.0
Flask-Cors==4.0.0
grpcio==1.60.0
grpcio-tools==1.60.0
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
MarkupSafe==2.1.3
mistune==3.0.2
packaging==23.2
protobuf==4.25.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.15.2
setuptools==69.0.2
six==1.16.0
typing_extensions==4.9.0
urllib3==2.1.0
validators==0.22.0
weaviate-client==4.4b2
Werkzeug==3.0.1
zipp==3.17.0
